# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Turnio API
  version: 0.1.0
  description: Contrato vivo entre el backend y el frontend de Turnio. Ver también
    /CONTRATO.md en la raíz del repo para convenciones que este schema no captura
    (auth, formato de errores, etc.).
paths:
...
  /api/caja/:
    get:
      operationId: caja_list
      description: |-
        La jornada de caja: histórico, apertura, egresos y cierre con arqueo.

        Leer (`list`/`retrieve`) exige `puede_cobrar` **o**
        `puede_ver_reportes` — cualquiera de las dos alcanza, porque quien
        opera la caja del día necesita mirar cierres pasados para cuadrar, y
        quien solo ve reportes también. Abrir, cerrar y registrar egresos
        exigen `puede_cobrar` sin excepción.

        Los **ingresos** no se crean acá: entran cobrando una venta
        (`POST /api/caja/ventas/{id}/cobrar/`). Es la regla central del
        módulo — la plata que entra siempre tiene una cuenta que la explica.
      parameters:
      - in: query
        name: fecha_desde
        schema:
          type: string
        description: Filtra desde esta fecha, inclusive (YYYY-MM-DD), sobre cuándo
          se abrió la caja.
      - in: query
        name: fecha_hasta
        schema:
          type: string
        description: Filtra hasta esta fecha, inclusive (YYYY-MM-DD), sobre cuándo
          se abrió la caja.
      tags:
      - caja
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CajaLista'
          description: ''
  /api/caja/{id}/:
    get:
      operationId: caja_retrieve
      description: |-
        La jornada de caja: histórico, apertura, egresos y cierre con arqueo.

        Leer (`list`/`retrieve`) exige `puede_cobrar` **o**
        `puede_ver_reportes` — cualquiera de las dos alcanza, porque quien
        opera la caja del día necesita mirar cierres pasados para cuadrar, y
        quien solo ve reportes también. Abrir, cerrar y registrar egresos
        exigen `puede_cobrar` sin excepción.

        Los **ingresos** no se crean acá: entran cobrando una venta
        (`POST /api/caja/ventas/{id}/cobrar/`). Es la regla central del
        módulo — la plata que entra siempre tiene una cuenta que la explica.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this caja.
        required: true
      tags:
      - caja
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CajaDetalle'
          description: ''
...
  /api/caja/ventas/{id}/cobrar/:
    post:
      operationId: caja_ventas_cobrar_create
      description: |-
        Cobra la venta, total o parcialmente. Crea el pago **y** su movimiento de caja en la misma transacción, así que exige caja abierta.

        Un pago mixto son dos llamadas a este endpoint sobre la misma venta, una por método. Cuando lo cobrado alcanza el total, la venta pasa a `pagada` y recién ahí se devengan las comisiones.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this venta.
        required: true
      tags:
      - caja
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cobrar'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Cobrar'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Cobrar'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venta'
          description: ''
...
components:
  schemas:
...
    CajaDetalle:
      type: object
      description: 'El detalle de una caja: sus movimientos y el resumen agregado.'
      properties:
        id:
          type: integer
          readOnly: true
        estado:
          allOf:
          - $ref: '#/components/schemas/CajaEstadoEnum'
          readOnly: true
        saldo_inicial:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          description: Efectivo con el que arranca el cajón (la base).
        abierta_por:
          type: integer
          readOnly: true
        abierta_por_nombre:
          type: string
          readOnly: true
        abierta_en:
          type: string
          format: date-time
          readOnly: true
        cerrada_por:
          type: integer
          readOnly: true
          nullable: true
        cerrada_por_nombre:
          type: string
          readOnly: true
          nullable: true
        cerrada_en:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        nota_cierre:
          type: string
          readOnly: true
        efectivo_esperado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: saldo_inicial + ingresos en efectivo − egresos en efectivo
            − devoluciones en efectivo, al momento del cierre.
        efectivo_contado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: Lo que la persona contó de verdad en el cajón.
        diferencia:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: efectivo_contado − efectivo_esperado. Negativo es faltante.
        movimientos:
          type: array
          items:
            $ref: '#/components/schemas/MovimientoCaja'
          readOnly: true
        resumen:
          allOf:
          - $ref: '#/components/schemas/ResumenCaja'
          readOnly: true
      required:
      - abierta_en
      - abierta_por
      - abierta_por_nombre
      - cerrada_en
      - cerrada_por
      - cerrada_por_nombre
      - diferencia
      - efectivo_contado
      - efectivo_esperado
      - estado
      - id
      - movimientos
      - nota_cierre
      - resumen
      - saldo_inicial
    CajaEstadoEnum:
      enum:
      - abierta
      - cerrada
      type: string
      description: |-
        * `abierta` - Abierta
        * `cerrada` - Cerrada
    CajaLista:
      type: object
      description: |-
        Una caja en el histórico — sin movimientos ni resumen, para que
        listar varias no cargue todo su detalle.
      properties:
        id:
          type: integer
          readOnly: true
        estado:
          allOf:
          - $ref: '#/components/schemas/CajaEstadoEnum'
          readOnly: true
        saldo_inicial:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          description: Efectivo con el que arranca el cajón (la base).
        abierta_por:
          type: integer
          readOnly: true
        abierta_por_nombre:
          type: string
          readOnly: true
        abierta_en:
          type: string
          format: date-time
          readOnly: true
        cerrada_por:
          type: integer
          readOnly: true
          nullable: true
        cerrada_por_nombre:
          type: string
          readOnly: true
          nullable: true
        cerrada_en:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        nota_cierre:
          type: string
          readOnly: true
        efectivo_esperado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: saldo_inicial + ingresos en efectivo − egresos en efectivo
            − devoluciones en efectivo, al momento del cierre.
        efectivo_contado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: Lo que la persona contó de verdad en el cajón.
        diferencia:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          description: efectivo_contado − efectivo_esperado. Negativo es faltante.
      required:
      - abierta_en
      - abierta_por
      - abierta_por_nombre
      - cerrada_en
      - cerrada_por
      - cerrada_por_nombre
      - diferencia
      - efectivo_contado
      - efectivo_esperado
      - estado
      - id
      - nota_cierre
      - saldo_inicial
...
    Cobrar:
      type: object
      description: |-
        Entrada de `POST /api/caja/ventas/{id}/cobrar/`.

        Un pago mixto son **dos llamadas** a este endpoint sobre la misma
        venta, una por método. Se prefirió eso a recibir una lista de pagos
        en un solo request porque cada pago es un hecho independiente que
        puede fallar por su cuenta (la caja se cerró entremedio, el monto
        excede el saldo), y un error parcial dentro de una lista es
        justamente lo que no se quiere en el dominio de dinero.
      properties:
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        metodo_pago:
          $ref: '#/components/schemas/MetodoPagoEnum'
      required:
      - metodo_pago
      - monto
...
    MetodoPagoEnum:
      enum:
      - efectivo
      - tarjeta
      - nequi
      - daviplata
      - bre_b
      - otro
      type: string
      description: |-
        * `efectivo` - Efectivo
        * `tarjeta` - Tarjeta
        * `nequi` - Nequi
        * `daviplata` - Daviplata
        * `bre_b` - Bre-B
        * `otro` - Otro
...
    MovimientoCaja:
      type: object
      description: |-
        Una línea del libro. **Solo lectura**: los movimientos no se crean
        sueltos ni se editan.

        Un ingreso nace de cobrar una venta (`POST /api/caja/ventas/{id}/cobrar/`),
        una devolución de devolver (`.../devolver/`), y un egreso tiene su
        propio endpoint (`POST /api/caja/egresos/`). No hay un "crear
        movimiento" genérico a propósito: era la puerta por la que entraba
        plata que ninguna venta explicaba.
      properties:
        id:
          type: integer
          readOnly: true
        caja:
          type: integer
          readOnly: true
        tipo:
          allOf:
          - $ref: '#/components/schemas/MovimientoCajaTipoEnum'
          readOnly: true
        metodo_pago:
          allOf:
          - $ref: '#/components/schemas/MetodoPagoEnum'
          readOnly: true
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          description: Siempre positivo. El signo lo da `tipo`, no el monto.
        concepto:
          type: string
          readOnly: true
        categoria:
          allOf:
          - $ref: '#/components/schemas/CategoriaEgresoEnum'
          readOnly: true
        venta:
          type: integer
          readOnly: true
          nullable: true
        registrado_por:
          type: integer
          readOnly: true
        registrado_por_nombre:
          type: string
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - caja
      - categoria
      - concepto
      - creado_en
      - id
      - metodo_pago
      - monto
      - registrado_por
      - registrado_por_nombre
      - tipo
      - venta
    MovimientoCajaTipoEnum:
      enum:
      - ingreso
      - egreso
      - devolucion
      type: string
      description: |-
        * `ingreso` - Ingreso
        * `egreso` - Egreso
        * `devolucion` - Devolución
...
    Pago:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        monto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        metodo_pago:
          allOf:
          - $ref: '#/components/schemas/MetodoPagoEnum'
          readOnly: true
        registrado_por:
          type: integer
          readOnly: true
        registrado_por_nombre:
          type: string
          readOnly: true
        movimiento:
          type: integer
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - creado_en
      - id
      - metodo_pago
      - monto
      - movimiento
      - registrado_por
      - registrado_por_nombre
...
    ResumenCaja:
      type: object
      description: |-
        Los totales de una caja, siempre calculados en caliente
        (`apps.caja.services.resumen_de`) — nunca persistidos.
      properties:
        total_ingresos:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        total_egresos:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        total_devoluciones:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        neto:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        por_metodo_pago:
          type: object
          additionalProperties:
            type: string
            format: decimal
            pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        egresos_por_categoria:
          type: object
          additionalProperties:
            type: string
            format: decimal
            pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        comisiones_por_empleado:
          type: array
          items:
            $ref: '#/components/schemas/ResumenComision'
        ventas_sin_cobrar:
          type: integer
          description: Cuántas ventas del negocio siguen pendientes o parciales. Informativo,
            no bloquea el cierre.
        saldo_inicial:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        ingresos_efectivo:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        egresos_efectivo:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        devoluciones_efectivo:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        efectivo_esperado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
      required:
      - comisiones_por_empleado
      - devoluciones_efectivo
      - efectivo_esperado
      - egresos_efectivo
      - egresos_por_categoria
      - ingresos_efectivo
      - neto
      - por_metodo_pago
      - saldo_inicial
      - total_devoluciones
      - total_egresos
      - total_ingresos
      - ventas_sin_cobrar
...
    Venta:
      type: object
      description: |-
        Una cuenta: qué se hizo, cuánto vale, cuánto se cobró.

        `estado`, `total` y todo lo relacionado con el cobro son de solo
        lectura: los fija el servidor a través de `cobrar`, `devolver` y
        `anular`. El cliente solo manda el cliente, los items y (si aplica)
        la evidencia.
      properties:
        id:
          type: integer
          readOnly: true
        cita:
          type: integer
          readOnly: true
          nullable: true
        nombre_cliente:
          type: string
          maxLength: 150
        telefono_cliente:
          type: string
          maxLength: 30
        items:
          type: array
          items:
            $ref: '#/components/schemas/VentaItem'
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        total_pagado:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        saldo_pendiente:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
        estado:
          allOf:
          - $ref: '#/components/schemas/VentaEstadoEnum'
          readOnly: true
        observaciones:
          type: string
        evidencia:
          type: string
          format: uri
          nullable: true
        pagos:
          type: array
          items:
            $ref: '#/components/schemas/Pago'
          readOnly: true
        devoluciones:
          type: array
          items:
            $ref: '#/components/schemas/Devolucion'
          readOnly: true
        creada_por:
          type: integer
          readOnly: true
        creada_por_nombre:
          type: string
          readOnly: true
        anulada_por:
          type: integer
          readOnly: true
          nullable: true
        anulada_en:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        motivo_anulacion:
          type: string
          readOnly: true
        creado_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - anulada_en
      - anulada_por
      - cita
      - creada_por
      - creada_por_nombre
      - creado_en
      - devoluciones
      - estado
      - id
      - items
      - motivo_anulacion
      - nombre_cliente
      - pagos
      - saldo_pendiente
      - total
      - total_pagado
    VentaEstadoEnum:
      enum:
      - pendiente
      - parcial
      - pagada
      - anulada
      type: string
      description: |-
        * `pendiente` - Pendiente de cobro
        * `parcial` - Pagada parcialmente
        * `pagada` - Pagada
        * `anulada` - Anulada
    VentaItem:
      type: object
      description: |-
        Una línea de la venta.

        Al **crear**, se manda `servicio` (y de ahí se copian descripción,
        precio y comisión) o bien `descripcion` + `precio_unitario` a mano,
        para vender algo que todavía no tiene catálogo. `empleado` es siempre
        obligatorio: es la fuente de verdad de la comisión.

        Al **leer**, `descripcion`, `precio_unitario` y `porcentaje_comision`
        son los valores congelados al momento de la venta, no los del
        catálogo de hoy.
      properties:
        id:
          type: integer
          readOnly: true
        servicio:
          type: integer
          nullable: true
        empleado:
          type: integer
        empleado_nombre:
          type: string
          readOnly: true
        descripcion:
          type: string
          maxLength: 200
        precio_unitario:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        cantidad:
          type: integer
          maximum: 2147483647
          minimum: 1
        porcentaje_comision:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          readOnly: true
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - empleado
      - empleado_nombre
      - id
      - porcentaje_comision
      - subtotal
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT